requirements.txt
# NLP & Document Processing
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
pdfplumber==0.10.2
python-docx==1.0.1
openpyxl==3.1.2

# Web Interface
streamlit==1.28.2

# LLM Integration
anthropic==0.5.0
tenacity==8.2.3

# Data Processing
pandas==2.1.1
numpy==1.24.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development Tools
black==23.9.1
isort==5.12.0
flake8==6.1.0
mypy==1.5.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
pydantic==2.4.2

# Optional Dependencies
PyMuPDF==1.23.7  # Alternative PDF processing if needed
tabula-py==2.9.0  # Additional table extraction capabilities
